requirements.txt
python-pptx==0.6.23
Pillow==10.2.0
